requirements.txt
django
djangorestframework
Python-dotenv
requests
psycopg2-binary
telebot
gunicorn
djangorestframework_simplejwt